refactor: 武器登録件数取得APIパッケージ配置変更
refs #008
