--- MobileSuitDatabase/src/main/java/jp/co/sunarch/mobilesuitDatabase/port/adapter/api/controller/internal/ArmsCountModel.java
@@ -1,4 +1,0 @@
-package jp.co.sunarch.mobilesuitDatabase.port.adapter.api.controller.internal;
-
-public record ArmsCountModel(Long count) {
-}
