+++ MobileSuitDatabase/src/main/java/jp/co/sunarch/mobilesuitDatabase/port/adapter/api/controller/internal/arms/ArmsCountModel.java
@@ -1,0 +1,4 @@
+package jp.co.sunarch.mobilesuitDatabase.port.adapter.api.controller.internal.arms;
+
+public record ArmsCountModel(Long count) {
+}
