+++ MobileSuitDatabase/src/main/java/jp/co/sunarch/mobilesuitDatabase/port/adapter/query/arms/ArmsQuery.java
@@ -2,7 +2,7 @@ package jp.co.sunarch.mobilesuitDatabase.port.adapter.query.arms;
 
 import java.util.List;
 
-import jp.co.sunarch.mobilesuitDatabase.port.adapter.api.controller.internal.ArmsCountModel;
+import jp.co.sunarch.mobilesuitDatabase.port.adapter.api.controller.internal.arms.ArmsCountModel;
 import jp.co.sunarch.mobilesuitDatabase.port.adapter.web.model.arms.ArmsModel;
 import lombok.Builder;
 import lombok.Value;
